requirements.txt
# Core Web Framework
fastapi==0.100.0
uvicorn[standard]==0.23.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.21
alembic==1.11.1

# HTTP Clients & API
httpx>=0.27.2
aiofiles>=24.1.0
requests==2.31.0

# Telegram Bot
python-telegram-bot==20.3

# OpenAI & AI
openai>=1.66.1
tiktoken==0.4.0

# Data Processing
pydantic>=2.10
pandas==2.0.3
numpy==1.24.3

# PDF & Images
reportlab==4.0.4
pillow~=10.4

# Visualization
matplotlib==3.7.1
networkx==3.1
plotly==5.15.0
pyvis==0.3.2

# Utilities
python-dotenv==1.0.0
pytz==2023.3
email-validator==2.0.0 